🚚 move: #49 圧縮処理のhooksのファイルを移動
