+++ src/app/post/create/_components/CreatePost/CreatePost.hooks.ts
@@ -6,7 +6,7 @@ import { useAtom } from 'jotai'
 import { useRouter } from 'next/navigation'
 import { useForm } from 'react-hook-form'
 import { z } from 'zod'
-import { useCompressImage } from '@/app/post/lib/compressImage'
+import { useCompressImage } from '@/app/_components/Form/hooks/useCompressImage'
 import { SelectedShop, selectedShopAtom } from '@/app/shop/atom'
 import { toast } from '@/components/ui/use-toast'
 import { handleApiError } from '@/libs/axios/handleError'
